migrated passwordconfirmation tests
