+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/PasswordConfirmation/tests/PasswordConfirmation.test.js
@@ -1,6 +1,6 @@
 // @flow
 import React from 'react';
-import {render, screen} from 'enzyme';
+import {fireEvent, render, screen} from '@testing-library/react';
 import debounce from 'debounce';
 import PasswordConfirmation from '../PasswordConfirmation';
 
@@ -18,47 +18,51 @@ test('Should render disabled input-components when disabled', () => {
     expect(debounce).toBeCalledWith(expect.any(Function), 500);
 });
 
-// test('Should only call onChange when both values match after the debounced time', () => {
-//     const changeSpy = jest.fn();
-//     const passwordConfirmation = shallow(<PasswordConfirmation onChange={changeSpy} />);
+test('Should only call onChange when both values match after the debounced time', () => {
+    const changeSpy = jest.fn();
+    render(<PasswordConfirmation onChange={changeSpy} />);
 
-//     expect(changeSpy).not.toBeCalled();
+    const inputs = screen.queryAllByDisplayValue('');
 
-//     passwordConfirmation.find('Input').at(0).simulate('change', 'asdf');
-//     passwordConfirmation.find('Input').at(1).simulate('change', 'jklö');
+    expect(changeSpy).not.toBeCalled();
 
-//     expect(changeSpy).not.toBeCalled();
+    fireEvent.change(inputs[0], {target: {value: 'asdf'}});
+    fireEvent.change(inputs[1], {target: {value: 'jklö'}});
 
-//     passwordConfirmation.find('Input').at(1).simulate('change', 'asdf');
-//     passwordConfirmation.find('Input').at(1).simulate('blur');
-//     expect(changeSpy).toBeCalledWith('asdf');
-// });
+    expect(changeSpy).not.toBeCalled();
 
-// test('Should mark the input fields as invalid if they do not match', () => {
-//     const changeSpy = jest.fn();
-//     const passwordConfirmation = shallow(<PasswordConfirmation onChange={changeSpy} />);
+    fireEvent.change(inputs[1], {target: {value: 'asdf'}});
+    fireEvent.blur(inputs[1]);
 
-//     expect(changeSpy).not.toBeCalled();
+    expect(changeSpy).toBeCalledWith('asdf');
+});
 
-//     passwordConfirmation.find('Input').at(0).simulate('change', 'asdf');
-//     passwordConfirmation.find('Input').at(1).simulate('change', 'jklö');
+test('Should mark the input fields as invalid if they do not match', () => {
+    const changeSpy = jest.fn();
+    const {container} = render(<PasswordConfirmation onChange={changeSpy} />);
 
-//     passwordConfirmation.find('Input').at(1).simulate('blur');
+    const inputs = screen.queryAllByDisplayValue('');
 
-//     expect(passwordConfirmation.find('Input').at(0).prop('valid')).toBe(false);
-//     expect(passwordConfirmation.find('Input').at(1).prop('valid')).toBe(false);
+    expect(changeSpy).not.toBeCalled();
 
-//     passwordConfirmation.find('Input').at(1).simulate('change', 'asdf');
-//     passwordConfirmation.find('Input').at(1).simulate('blur');
+    fireEvent.change(inputs[0], {target: {value: 'asdf'}});
+    fireEvent.change(inputs[1], {target: {value: 'jklö'}});
+    fireEvent.blur(inputs[1]);
 
-//     expect(passwordConfirmation.find('Input').at(0).prop('valid')).toBe(true);
-//     expect(passwordConfirmation.find('Input').at(1).prop('valid')).toBe(true);
-// });
+    // eslint-disable-next-line testing-library/no-container, testing-library/no-node-access
+    expect(container.querySelector('.error')).toBeInTheDocument();
 
-// test('Should mark the input fields as invalid if the valid prop is false', () => {
-//     const changeSpy = jest.fn();
-//     const passwordConfirmation = shallow(<PasswordConfirmation onChange={changeSpy} valid={false} />);
+    fireEvent.change(inputs[1], {target: {value: 'asdf'}});
+    fireEvent.blur(inputs[1]);
 
-//     expect(passwordConfirmation.find('Input').at(0).prop('valid')).toBe(false);
-//     expect(passwordConfirmation.find('Input').at(1).prop('valid')).toBe(false);
-// });
+    // eslint-disable-next-line testing-library/no-container, testing-library/no-node-access
+    expect(container.querySelector('.error')).not.toBeInTheDocument();
+});
+
+test('Should mark the input fields as invalid if the valid prop is false', () => {
+    const changeSpy = jest.fn();
+    const {container} = render(<PasswordConfirmation onChange={changeSpy} valid={false} />);
+
+    // eslint-disable-next-line testing-library/no-container, testing-library/no-node-access
+    expect(container.querySelector('.error')).toBeInTheDocument();
+});
